Module Sc_core.Printer

Pretty-printing utilities

val pp_generic_preprocessing_error : pp_operation:(Stdlib.Format.formatter -> unit) -> pp_error:'a Fmt.t -> 'a Types.generic_preprocessing_error Fmt.t

pp_generic_preprocessing_error ~pp_operation ~pp_error ppf error prints using ppf a message about a failed pre-processing operation; pp_operation should give a short description of what failed, and pp_error print details about the error itself.